pyproject.toml
[project]
name = "vectorsearch"
version = "0.1.0"
description = "Vector store per ricerche di similarity con LangChain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.1.0",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.0",
    "langchain-text-splitters>=1.0.0",
    "python-dotenv>=1.2.1",
    "unstructured[md]>=0.16.0",
]

[project.scripts]
vs = "vectorsearch.cli:main"
rag = "vectorsearch.rag:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vectorsearch"]

[dependency-groups]
dev = [
    "hatchling>=1.28.0",
]
